% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatDiffusion.R
\name{diffusionMat}
\alias{diffusionMat}
\title{Calculate the heat diffusion matrix when reaching equilibrium}
\usage{
diffusionMat(links = interactions_brain, directed = FALSE, beta = 0.55)
}
\arguments{
\item{links}{A two-column matrix or data frame, with each row representing one interaction}

\item{directed}{Whether the interactions are directed}

\item{beta}{Insulating parameter between 0 and 1}
}
\value{
The heat diffusion matrix
}
\description{
This function calculate the equilibrium heat diffusion
matrix, using HotNet2 algorithm (Leiserson
MDM, et al. Nat Genet. 2015).
}
