% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatDiffusion.R
\name{equilibriumHeat}
\alias{equilibriumHeat}
\title{Calculate heat after reaching equilibrium}
\usage{
equilibriumHeat(diffusion, heat)
}
\arguments{
\item{diffusion}{The diffusion matrix}

\item{heat}{A numeric vector or one-column matrix representing vertices' initial heat}
}
\value{
A numeric vector representing equilibrium heat of vectices
}
\description{
This function calculate the equilibrium heat of each
vertex in the network.
}
