% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatDiffusion.R
\name{fillHeat}
\alias{fillHeat}
\title{Fill in the missing heat}
\usage{
fillHeat(links = interactions_brain, heat = vertices_heat, fill = "mean",
  ...)
}
\arguments{
\item{links}{A two-column matrix or data frame, with each row representing one interaction}

\item{heat}{A numeric vector representing initial heat of each vectex in the network}

\item{fill}{A numeric value, character, or function to specialize the way to fill in missing values}
}
\value{
A numeric vector representing heat of vertices
}
\description{
This function finds the vertices in the interaction network
whose initial heat is not given, and fill with heat using
specific value or function.
}
