% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{heatModules}
\alias{heatModules}
\title{The automatic procedure to detect modules in the interaction network}
\usage{
heatModules(links = interactions_brain, heat = NULL, force_HD = FALSE,
  fill = "mean", directed = FALSE, beta = 0.55, minSize = 20, ...)
}
\arguments{
\item{links}{A two-column matrix or data frame, with each row representing one interaction}

\item{heat}{A numeric vector representing initial heat of each vectex in the network}

\item{force_HD}{If there are no more than half of the vertices have known heat, whether the heat diffusion procedure should be applied}

\item{fill}{A numeric value, character, or function to specialize the way to fill in missing values}

\item{directed}{Whether the interactions are directed}

\item{beta}{Insulating parameter between 0 and 1}

\item{minSize}{The minimum requirement of identified modules}

\item{...}{Other parameters for 'dynamicTreeCut' in TOM-based module detection}
}
\value{
A list with each element representing vertices of one module
}
\description{
This function integrates all steps including missing value
imputation, heat diffusion and module detection into an
automatic pipeline.
}
