% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatDiffusion.R
\name{initialHeat}
\alias{initialHeat}
\title{Generate heat based on log2fc and p-value}
\usage{
initialHeat(fc = NULL, p = NULL)
}
\arguments{
\item{fc}{A numeric vector indicating fold change of expression}

\item{p}{A numeric vector indicating P-value of respective test}
}
\value{
A numeric vector of intial heat with the same length as fc and p-value
}
\description{
This function generate the initial heat for genes with the
given fold change and p-value vector, as the square root of
Log2-transformed FC times -Log10-transformed P-value.
}
