% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{TOMModules}
\alias{TOMModules}
\title{Module detection based on Topological Overlap Matrix}
\usage{
TOMModules(graph, minSize = 20, ...)
}
\arguments{
\item{graph}{The igraph object representing the interaction network}

\item{minSize}{The minimum requirement of identified modules}
}
\value{
A list with each element representing vertices of one module
}
\description{
This function calculates TOM matrix using the implementation
in WGCNA package, and applied hierarchical clustering to identify
modules.
}
