% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{weightedGraph}
\alias{weightedGraph}
\title{Generate the weighted igraph object}
\usage{
weightedGraph(links = interactions_brain, directed = FALSE, heat = NULL)
}
\arguments{
\item{links}{A two-column matrix or data frame, with each row representing one interaction}

\item{directed}{Whether the interactions are directed}

\item{heat}{A numeric vector representing vertices' heat}
}
\value{
A igraph object representing the interaction network. Edges are weighted accordingly
}
\description{
This function generate the igraph object based on the given
interactions. The weight of each edge is determined as the
heat differences between the two interacting vertices.
}
